change binmat description in glossary (#260)
* Correct correcting of styling 2

"hackmud card game" is the title of something which would not be italicized. If the goal is to add hackmud as a title, rearranging the sentence a hair let's us do that in a way that is more correct and less visually confusing.

* fix linter

Accidental removal of new line. This kills the linter.

* Fix fix linter

Accidental removal of new line. This kills the linter, again.